[feat] add `initialize_tool_auth`, `wait_tool_auth` in pocket
