+++ libs/extensions/llamaindex/hyperpocket_llamaindex/pocket_llamaindex.py
@@ -13,7 +13,7 @@ from hyperpocket.tool import Tool
 
 class PocketLlamaindex(Pocket):
     def get_tools(self) -> List[BaseTool]:
-        tools = [self.get_tool(pk) for pk in self.tools.values()]
+        tools = [self.get_tool(pk) for pk in self.core.tools.values()]
         return tools
 
     def get_tool(self, pocket_tool: Tool) -> BaseTool:
