+++ libs/extensions/openai/hyperpocket_openai/pocket_openai.py
@@ -34,7 +34,7 @@ class PocketOpenAI(Pocket):
 
     def get_open_ai_tool_specs(self) -> List[dict]:
         specs = []
-        for tool in self.tools.values():
+        for tool in self.core.tools.values():
             spec = self.get_open_ai_tool_spec(tool)
             specs.append(spec)
         return specs
