+++ libs/hyperpocket/hyperpocket/pocket_core.py
@@ -1,0 +1,210 @@
+import asyncio
+import pathlib
+from typing import Any, Optional, Callable
+
+from hyperpocket.builtin import get_builtin_tools
+from hyperpocket.config import pocket_logger
+from hyperpocket.pocket_auth import PocketAuth
+from hyperpocket.repository import Lockfile
+from hyperpocket.repository.lock import LocalLock, GitLock
+from hyperpocket.tool import Tool, ToolRequest
+from hyperpocket.tool.function import from_func
+from hyperpocket.tool.wasm import WasmTool
+from hyperpocket.tool.wasm.tool import WasmToolRequest
+from hyperpocket.tool_like import ToolLike
+
+
+class PocketCore:
+    auth: PocketAuth
+    tools: dict[str, Tool]
+
+    def __init__(self,
+                 tools: list[ToolLike],
+                 auth: PocketAuth = None,
+                 lockfile_path: Optional[str] = None,
+                 force_update: bool = False):
+        if auth is None:
+            auth = PocketAuth()
+        self.auth = auth
+
+        if lockfile_path is None:
+            lockfile_path = "./pocket.lock"
+        lockfile_pathlib_path = pathlib.Path(lockfile_path)
+        lockfile = Lockfile(lockfile_pathlib_path)
+        tool_likes = []
+        for tool_like in tools:
+            if isinstance(tool_like, str):
+                if pathlib.Path(tool_like).exists():
+                    lock = LocalLock(tool_like)
+                    req = WasmToolRequest(lock, "")
+                else:
+                    lock = GitLock(repository_url=tool_like, git_ref='HEAD')
+                    req = WasmToolRequest(lock, "")
+                lockfile.add_lock(lock)
+                tool_likes.append(req)
+            elif isinstance(tool_like, WasmToolRequest):
+                lockfile.add_lock(tool_like.lock)
+                tool_likes.append(tool_like)
+            elif isinstance(tool_like, ToolRequest):
+                raise ValueError(f"unreachable. tool_like:{tool_like}")
+            elif isinstance(tool_like, WasmTool):
+                raise ValueError("WasmTool should pass ToolRequest instance instead.")
+            else:
+                tool_likes.append(tool_like)
+        lockfile.sync(force_update=force_update, referenced_only=True)
+
+        self.tools = dict()
+        for tool_like in tool_likes:
+            tool = self._load_tool(tool_like, lockfile)
+            if tool.name in self.tools:
+                pocket_logger.error(f"Duplicate tool name: {tool.name}.")
+                raise ValueError(f"Duplicate tool name: {tool.name}")
+            self.tools[tool.name] = tool
+
+        pocket_logger.info(f"All Registered Tools Loaded successfully. total registered tools : {len(self.tools)}")
+
+        builtin_tools = get_builtin_tools(self.auth)
+        for tool in builtin_tools:
+            self.tools[tool.name] = tool
+        pocket_logger.info(f"All BuiltIn Tools Loaded successfully. total tools : {len(self.tools)}")
+
+    async def acall(self,
+                    tool_name: str,
+                    body: Any,
+                    thread_id: str = 'default',
+                    profile: str = 'default',
+                    *args, **kwargs) -> tuple[str, bool]:
+        """
+        Invoke tool asynchronously, not that different from `Pocket.invoke`
+        But this method is called only in subprocess.
+
+        This function performs the following steps:
+            1. `prepare_auth` : preparing the authentication process for the tool if necessary.
+            2. `authenticate` : performing authentication that needs to invoke tool.
+            3. `tool_call` : Executing tool actually with authentication information.
+
+        Args:
+            tool_name(str): tool name to invoke
+            body(Any): tool arguments. should be json format
+            thread_id(str): thread id
+            profile(str): profile name
+
+        Returns:
+            tuple[str, bool]: tool result and state.
+        """
+        tool = self._tool_instance(tool_name)
+        if tool.auth is not None:
+            callback_info = self.prepare_auth(tool_name, thread_id, profile, **kwargs)
+            if callback_info:
+                return callback_info, True
+        # 02. authenticate
+        credentials = await self.authenticate(tool_name, thread_id, profile, **kwargs)
+        # 03. call tool
+        result = await self.tool_call(tool_name, body=body, envs=credentials, **kwargs)
+        return result, False
+
+    def prepare_auth(self,
+                     tool_name: str,
+                     thread_id: str = 'default',
+                     profile: str = 'default',
+                     **kwargs) -> Optional[str]:
+        """
+        Prepares the authentication process for the tool if necessary.
+        Returns callback URL and whether the tool requires authentication.
+
+        Args:
+            tool_name(str): tool name to invoke
+            thread_id(str): thread id
+            profile(str): profile name
+
+        Returns:
+            Optional[str]: callback URI if necessary
+        """
+        tool = self._tool_instance(tool_name)
+        if tool.auth is None:
+            return None
+
+        auth_req = self.auth.make_request(
+            auth_handler_name=tool.auth.auth_handler,
+            auth_provider=tool.auth.auth_provider,
+            auth_scopes=tool.auth.scopes)
+
+        return self.auth.prepare(
+            auth_req=auth_req,
+            auth_handler_name=tool.auth.auth_handler,
+            auth_provider=tool.auth.auth_provider,
+            thread_id=thread_id,
+            profile=profile,
+            **kwargs
+        )
+
+    async def authenticate(
+            self,
+            tool_name: str,
+            thread_id: str = 'default',
+            profile: str = 'default',
+            **kwargs) -> dict[str, str]:
+        """
+        Authenticates the handler included in the tool and returns credentials.
+
+        Args:
+            tool_name(str): tool name to invoke
+            thread_id(str): thread id
+            profile(str): profile name
+
+        Returns:
+            dict[str, str]: credentials
+        """
+        tool = self._tool_instance(tool_name)
+        if tool.auth is None:
+            return {}
+        auth_req = self.auth.make_request(
+            auth_handler_name=tool.auth.auth_handler,
+            auth_provider=tool.auth.auth_provider,
+            auth_scopes=tool.auth.scopes)
+        auth_ctx = await self.auth.authenticate_async(
+            auth_req=auth_req,
+            auth_handler_name=tool.auth.auth_handler,
+            auth_provider=tool.auth.auth_provider,
+            thread_id=thread_id,
+            profile=profile,
+            **kwargs,
+        )
+        return auth_ctx.to_dict()
+
+    async def tool_call(self, tool_name: str, **kwargs) -> str:
+        """
+        Executing tool actually
+
+        Args:
+            tool_name(str): tool name to invoke
+            kwargs(dict): keyword arguments. authentication information is passed through this.
+
+        Returns:
+            str: tool result
+        """
+        tool = self._tool_instance(tool_name)
+        try:
+            return await asyncio.wait_for(tool.ainvoke(**kwargs), timeout=180)
+        except asyncio.TimeoutError:
+            pocket_logger.warning("Timeout tool call.")
+            return "timeout tool call"
+
+    def _tool_instance(self, tool_name: str) -> Tool:
+        return self.tools[tool_name]
+
+    @staticmethod
+    def _load_tool(tool_like: ToolLike, lockfile: Lockfile) -> Tool:
+        pocket_logger.info(f"Loading Tool {tool_like}")
+        if isinstance(tool_like, Tool):
+            tool = tool_like
+        elif isinstance(tool_like, ToolRequest):
+            tool = Tool.from_tool_request(tool_like, lockfile=lockfile)
+        elif isinstance(tool_like, Callable):
+            tool = from_func(tool_like)
+        else:
+            raise ValueError(f"Invalid tool type: {type(tool_like)}")
+
+        pocket_logger.info(f"Complete Loading Tool {tool.name}")
+        return tool
+
