+++ libs/hyperpocket/hyperpocket/tool_like.py
@@ -1,0 +1,5 @@
+from typing import Union, Callable
+
+from hyperpocket.tool import ToolRequest, Tool
+
+ToolLike = Union[Tool, str, Callable, ToolRequest]
